+++ src/decoder/image/ppm_decoder.ml
@@ -20,8 +20,6 @@
 
  *****************************************************************************)
 
-module Img = Image.RGBA32
-
 let log = Log.make ["decoder";"ppm"]
 
 let load_image fname =
@@ -30,7 +28,7 @@ let load_image fname =
   let data = Bytes.create len in
   really_input ic data 0 len;
   close_in ic;
-  Img.of_PPM (Bytes.unsafe_to_string data)
+  Image.YUV420.of_PPM (Bytes.unsafe_to_string data)
 
 let () =
   Decoder.image_file_decoders#register "ppm"