+++ src/decoder/ogg_decoder.ml
@@ -22,9 +22,7 @@
 
 (** Decode and read ogg files. *)
 
-module Img = Image.RGBA32
-module Gen = Image.Generic
-module P = Gen.Pixel
+module P = Image.Generic.Pixel
 
 let log = Log.make ["decoder";"ogg"]
 
@@ -47,31 +45,39 @@ let converter () =
         let converter =
           Video_converter.find_converter
              (P.YUV format)
-             (P.RGB P.RGBA32)
+             (P.YUV P.YUVJ420)
         in
         current_format := Some (format,converter) ;
         converter)
 
-(** Convert a video frame to RGB *)
-let video_convert =
+(** Convert a video frame to YUV *)
+let video_convert scale =
   let converter = converter () in
   (fun buf ->
-    let converter = converter buf.Ogg_demuxer.format in
     let width = Lazy.force Frame.video_width in
-    let height = Lazy.force Frame.video_height in
-    let rgb = Img.create width height in
-    let frame = Gen.of_RGBA32 rgb in
-    let sframe = 
-      Image.YUV420.make 
-        buf.Ogg_demuxer.frame_width buf.Ogg_demuxer.frame_height 
-        buf.Ogg_demuxer.y buf.Ogg_demuxer.y_stride 
-        buf.Ogg_demuxer.u buf.Ogg_demuxer.v 
-        buf.Ogg_demuxer.uv_stride 
-   in
-    converter
-      (Gen.of_YUV420 sframe)
-      frame;
-    rgb)
+      let height = Lazy.force Frame.video_height in
+    if buf.Ogg_demuxer.format <> Ogg_demuxer.Yuvj_420 then
+      let img =
+        Image.YUV420.make
+          buf.Ogg_demuxer.frame_width buf.Ogg_demuxer.frame_height
+          buf.Ogg_demuxer.y buf.Ogg_demuxer.y_stride
+          buf.Ogg_demuxer.u buf.Ogg_demuxer.v buf.Ogg_demuxer.uv_stride
+      in
+      let img2 = Video.Image.create width height in
+      scale img img2;
+      img2
+    else
+      let converter = converter buf.Ogg_demuxer.format in
+      let yuv = Video.Image.create width height in
+      let frame = Image.Generic.of_YUV420 yuv in
+      let sframe =
+        Image.YUV420.make
+          buf.Ogg_demuxer.frame_width buf.Ogg_demuxer.frame_height
+          buf.Ogg_demuxer.y buf.Ogg_demuxer.y_stride
+          buf.Ogg_demuxer.u buf.Ogg_demuxer.v buf.Ogg_demuxer.uv_stride
+      in
+      converter (Image.Generic.of_YUV420 sframe) frame;
+      yuv)
 
 (** Stupid nearest neighbour resampling.
   * For meaningful results, one should first partially apply the freq params,
@@ -141,6 +147,7 @@ let create_decoder ?(merge_tracks=false) source mode input =
   in
   let audio_resample = Rutils.create_audio () in
   let video_resample = video_resample () in
+  let video_scale = Video_converter.scaler () ~proportional:true in
   let decode_audio = mode = `Both || mode = `Audio in
   let decode_video = mode = `Both || mode = `Video in
   let started = ref false in
@@ -226,14 +233,14 @@ let create_decoder ?(merge_tracks=false) source mode input =
             decoder track 
         in
         let out_freq = Lazy.force Frame.video_rate in
-        let rgb = video_convert buf in
+        let rgb = video_convert video_scale buf in
         let in_freq = float info.Ogg_demuxer.fps_numerator /. 
                       float info.Ogg_demuxer.fps_denominator
         in
         let stream = 
-          video_resample ~in_freq ~out_freq [|rgb|] 0 1 
+          video_resample ~in_freq ~out_freq (Video.single rgb) 0 1
         in
-        Generator.put_video buffer [|stream|] 0 (Array.length stream) ;
+        Generator.put_video buffer [|stream|] 0 (Video.length stream) ;
       in
       let decode_audio, decode_video =
         if decode_audio && decode_video then