+++ src/operators/frei0r_op.ml
@@ -23,8 +23,6 @@
 open Source
 open Extralib
 
-module Img = Image.RGBA32
-
 type t = Float | Int | Bool
 
 let log = Log.make ["frei0r"]
@@ -64,12 +62,16 @@ object
       params ();
       let rgb = rgb.(0) in
       for i = offset to offset + length - 1 do
-        let img = rgb.(i) in
-        if bgra then Img.swap_rb img;
-        let src = Img.data (Img.copy img) in
-        let dst = Img.data img in
+        (* TODO: we could try to be more efficient than converting to/from RGBA32 and swap colors... *)
+        let img = Video.get rgb i in
+        let img = Image.YUV420.to_RGBA32 img in
+        if bgra then Image.RGBA32.swap_rb img;
+        let src = Image.RGBA32.data (Image.RGBA32.copy img) in
+        let dst = Image.RGBA32.data img in
         Frei0r.update1 instance t src dst;
-        if bgra then Img.swap_rb img;
+        if bgra then Image.RGBA32.swap_rb img;
+        let img = Image.YUV420.of_RGBA32 img in
+        Video.set rgb i img;
         t <- t +. dt
       done
 end
@@ -119,15 +121,20 @@ object
       let rgb = rgb.(0) in
       let rgb' = rgb'.(0) in
       for i = offset to offset + length - 1 do
-        let img = rgb.(i) in
-        let img' = rgb'.(i) in
-        if bgra then Img.swap_rb img;
-        if bgra then Img.swap_rb img';
-        let src = Img.data (Img.copy img) in
-        let src' = Img.data img' in
-        let dst = Img.data img in
+        (* TODO: we could try to be more efficient than converting to/from RGBA32 and swap colors... *)
+        let img = Video.get rgb i in
+        let img = Image.YUV420.to_RGBA32 img in
+        let img' = Video.get rgb' i in
+        let img' = Image.YUV420.to_RGBA32 img' in
+        if bgra then Image.RGBA32.swap_rb img;
+        if bgra then Image.RGBA32.swap_rb img';
+        let src = Image.RGBA32.data (Image.RGBA32.copy img) in
+        let src' = Image.RGBA32.data img' in
+        let dst = Image.RGBA32.data img in
         Frei0r.update2 instance t src src' dst;
-        if bgra then Img.swap_rb img;
+        if bgra then Image.RGBA32.swap_rb img;
+        let img = Image.YUV420.of_RGBA32 img in
+        Video.set rgb i img;
         t <- t +. dt
       done
     | _ -> ()
@@ -159,11 +166,14 @@ object
       let buf = VFrame.content_of_type frame ~channels:1 in
       let buf = buf.(0) in
       for i = start to stop - 1 do
+        failwith "TODO: restore frei0r";
+        (*
         let img = buf.(i) in
         let buf = Img.data img in
         Frei0r.update0 instance t buf;
         if bgra then Img.swap_rb img;
         t <- t +. dt
+         *)
       done;
       VFrame.add_break frame stop
     end